Fix redirects and functionality.
Change adding options to questions to work in modals.

Add checks to prevent setting more than one option as correct.

Add sparate views for deleting and editing questions and options so that they redirect to quiz or questions depending on where the detail view was accessed from.

Fix adding images to quiz.